Added RK1 and UnitCellShift stability comps to MultiLatticeKIM

+++ LatticeStatics/Utility/UtilityFunctions.h
@@ -19,8 +19,8 @@ using namespace std;
 char kbhitWait();
 int EnterDebugMode();
 
-int FullScanRank1Convex3D(CBKinematics const* const CBK, Matrix const& K, double const& dx);
-int FullScanRank1Convex2D(Matrix const& K, double const& dx);
+double FullScanRank1Convex3D(CBKinematics const* const CBK, Matrix const& K, double const& dx);
+double FullScanRank1Convex2D(Matrix const& K, double const& dx);
 int Rank1Convex3D(CBKinematics const* const CBK, Matrix const& K, double const& dx);
 int Rank1Convex2D(Matrix const& K, double const& dx);
